refactor: minor cleanup in selectors
