+++ src/chart_types/partition_chart/state/selectors/tooltip.ts
@@ -21,9 +21,9 @@ import createCachedSelector from 're-reselect';
 
 import { TooltipInfo } from '../../../../components/tooltip/types';
 import { percentValueGetter, sumValueGetter } from '../../layout/config';
+import { valueGetterFunction } from '../../layout/viewmodel/scenegraph';
 import { getPartitionSpec } from './partition_spec';
 import { getPickedShapes } from './picked_shapes';
-import { valueGetterFunction } from './scenegraph';
 
 const EMPTY_TOOLTIP = Object.freeze({
   header: null,
